upgrade to bytecode V1_7
The bytecode could not pass java verification as the use/abuse of local
varibles, the code.getLocal/freeLocal destroys the stack frame map, as
one local variable could change it's type several times in a method.

In this commit, a lot of the local variable allocations are replaced
with stack slots, but there are still some cases needs fix.